Move actual methods into EntityMetadata and refactor ETPVI to use the presenter.
Tests also refactored, but only partial functionality move (only
versions service).  Unknown whether this is functional or not, but at
least it compiles.
Still to do: Make it work, then add the functionality that creates the
Governance and permissions things.
